data_loader/requirements.txt
aiohttp==3.8.3
asyncpg==0.27.0
python-dotenv==0.20.0
python-dateutil==2.8.2
Faker==17.6.0
